requirements.txt
gradio
Pillow
rembg
torch
torchvision
numpy
transformers
